+++ asphalt/core/command.py
@@ -1,10 +1,10 @@
+from typing import Optional
+
 import click
 import yaml
 
 from asphalt.core.component import component_types
-from asphalt.core.util import PluginContainer
-
-runners = PluginContainer('asphalt.core.runners')
+from asphalt.core.runner import run_application, policies
 
 
 @click.group()
@@ -13,16 +13,19 @@ def main():
 
 
 @main.command(help='Read a configuration file and start the application.')
-@click.argument('config', type=click.File())
-@click.option('--unsafe', is_flag=True,
+@click.argument('configfile', type=click.File())
+@click.option('--unsafe', is_flag=True, default=False,
               help='use unsafe mode when loading YAML (enables markup extensions)')
-def run(config, unsafe: bool=False):
+@click.option('-l', '--loop', type=click.Choice(policies.names),
+              help='alternate event loop policy')
+def run(configfile, unsafe: bool, loop: Optional[str]):
     # Read the configuration from the supplied YAML file
-    config_data = yaml.load(config) if unsafe else yaml.safe_load(config)
+    config_data = yaml.load(configfile) if unsafe else yaml.safe_load(configfile)
     assert isinstance(config_data, dict), 'the document root element must be a dictionary'
 
-    # Get a reference to the runner
-    runner = runners.resolve(config_data.pop('runner', 'asyncio'))
+    # Override the event loop policy if specified
+    if loop:
+        config_data['event_loop_policy'] = loop
 
     # Instantiate the root component
     try:
@@ -33,7 +36,7 @@ def run(config, unsafe: bool=False):
         component = component_types.create_object(**component_config)
 
     # Start the application
-    runner(component, **config_data)
+    run_application(component, **config_data)
 
 
 if __name__ == '__main__':  # pragma: no cover