+++ tests/test_command.py
@@ -7,16 +7,15 @@ from click.testing import CliRunner
 from asphalt.core import command
 from asphalt.core.component import ContainerComponent
 
-mock_runner = None
-
 
 @pytest.fixture
 def runner():
     return CliRunner()
 
 
+@pytest.mark.parametrize('loop', [None, 'uvloop'], ids=['default', 'override'])
 @pytest.mark.parametrize('unsafe', [False, True], ids=['safe', 'unsafe'])
-def test_run(runner, unsafe):
+def test_run(runner, unsafe, loop):
     if unsafe:
         component_class = '!!python/name:{0.__module__}.{0.__name__}'.format(ContainerComponent)
     else:
@@ -24,7 +23,7 @@ def test_run(runner, unsafe):
 
     config = """\
 ---
-runner: test_command:mock_runner
+event_loop_policy: bogus
 component:
   type: {}
 logging:
@@ -34,17 +33,22 @@ logging:
     args = ['test.yml']
     if unsafe:
         args.append('--unsafe')
+    if loop:
+        args.extend(['--loop', loop])
 
-    with runner.isolated_filesystem(), patch('test_command.mock_runner') as mock_runner:
+    with runner.isolated_filesystem(), patch('asphalt.core.command.run_application') as run_app:
         Path('test.yml').write_text(config)
         result = runner.invoke(command.run, args)
 
         assert result.exit_code == 0
-        assert mock_runner.call_count == 1
-        args, kwargs = mock_runner.call_args
+        assert run_app.call_count == 1
+        args, kwargs = run_app.call_args
+        assert len(args) == 1
         assert isinstance(args[0], ContainerComponent)
-        assert len(kwargs) == 1
-        assert kwargs['logging'] == {'version': 1, 'disable_existing_loggers': False}
+        assert kwargs == {
+            'event_loop_policy': loop or 'bogus',
+            'logging': {'version': 1, 'disable_existing_loggers': False}
+        }
 
 
 def test_run_missing_component_key(tmpdir, runner):